requirements.txt
asgiref==3.11.1
Django==5.2.12
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.15
inflection==0.5.1
packaging==26.0
PyJWT==2.12.1
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
